+++ src/com/koxudaxi/pydantic/Pydantic.kt
@@ -27,6 +27,7 @@ import com.jetbrains.python.statistics.modules
 import java.util.regex.Pattern
 
 const val BASE_MODEL_Q_NAME = "pydantic.main.BaseModel"
+const val ROOT_MODEL_Q_NAME = "pydantic.root_model.RootModel"
 const val GENERIC_MODEL_Q_NAME = "pydantic.generics.GenericModel"
 const val DATA_CLASS_Q_NAME = "pydantic.dataclasses.dataclass"
 const val DATA_CLASS_SHORT_Q_NAME = "pydantic.dataclass"
@@ -140,6 +141,10 @@ val V2_VALIDATOR_QUALIFIED_NAMES = listOf(
     MODEL_VALIDATOR_SHORT_QUALIFIED_NAME
 )
 
+val MODEL_VALIDATOR_QUALIFIED_NAMES = listOf(
+    MODEL_VALIDATOR_QUALIFIED_NAME,
+    MODEL_VALIDATOR_SHORT_QUALIFIED_NAME
+)
 val FIELD_VALIDATOR_Q_NAMES = listOf(
     VALIDATOR_Q_NAME,
     VALIDATOR_SHORT_Q_NAME,
@@ -230,6 +235,10 @@ internal fun isSubClassOfPydanticBaseModel(pyClass: PyClass, context: TypeEvalCo
     return pyClass.isSubclass(BASE_MODEL_Q_NAME, context)
 }
 
+internal fun isSubClassOfPydanticRootModel(pyClass: PyClass, context: TypeEvalContext): Boolean {
+    return pyClass.isSubclass(ROOT_MODEL_Q_NAME, context)
+}
+
 internal fun isSubClassOfBaseSetting(pyClass: PyClass, context: TypeEvalContext): Boolean {
     return pyClass.isSubclass(BASE_SETTINGS_Q_NAME, context)
 }
@@ -241,13 +250,9 @@ internal fun isSubClassOfCustomBaseModel(pyClass: PyClass, context: TypeEvalCont
 internal val PyClass.isBaseSettings: Boolean get() = qualifiedName == BASE_SETTINGS_Q_NAME
 
 
-internal fun hasDecorator(pyDecoratable: PyDecoratable, refNames: List<QualifiedName>): Boolean {
-    return pyDecoratable.decoratorList?.decorators?.mapNotNull { it.callee as? PyReferenceExpression }?.any {
-        PyResolveUtil.resolveImportedElementQNameLocally(it).any { decoratorQualifiedName ->
-            refNames.any { refName -> decoratorQualifiedName == refName }
-        }
-    } ?: false
-}
+internal fun hasDecorator(pyDecoratable: PyDecoratable, refNames: List<QualifiedName>): Boolean =
+    pyDecoratable.decoratorList?.decorators?.any {it.include(refNames)} ?: false
+
 
 internal val PyClass.isPydanticDataclass: Boolean get() = hasDecorator(this, DATA_CLASS_QUALIFIED_NAMES)
 
@@ -269,11 +274,27 @@ internal fun isDataclassMissing(pyTargetExpression: PyTargetExpression): Boolean
     return pyTargetExpression.qualifiedName == DATACLASS_MISSING
 }
 
-internal fun PyFunction.isValidatorMethod(pydanticVersion: KotlinVersion?): Boolean =
+internal fun PyFunction.hasValidatorMethod(pydanticVersion: KotlinVersion?): Boolean =
     hasDecorator(this, if(pydanticVersion.isV2) V2_VALIDATOR_QUALIFIED_NAMES else VALIDATOR_QUALIFIED_NAMES)
 
+internal fun PyDecorator.include(refNames: List<QualifiedName>): Boolean = (callee as? PyReferenceExpression)?.let {
+        PyResolveUtil.resolveImportedElementQNameLocally(it).any { decoratorQualifiedName ->
+            refNames.any { refName -> decoratorQualifiedName == refName }
+        }
+} ?: false
 
+internal val PyKeywordArgument.value: PyExpression?
+    get() = when (val value = valueExpression) {
+            is PyReferenceExpression -> (value.reference.resolve() as? PyTargetExpression)?.findAssignedValue()
+            else -> value
+        }
 
+internal fun PyFunction.hasModelValidatorModeAfter(): Boolean = decoratorList?.decorators
+    ?.filter { it.include(MODEL_VALIDATOR_QUALIFIED_NAMES) }
+    ?.any { modelValidator ->
+        modelValidator.argumentList?.getKeywordArgument("mode")
+            ?.let { it.value as? PyStringLiteralExpression }?.stringValue == "after"
+    } ?: false
 internal val PyClass.isConfigClass: Boolean get() = name == "Config"
 
 
@@ -558,6 +579,10 @@ fun getFieldName(
     return when (pydanticVersion?.major) {
         0 -> when {
             config["allow_population_by_alias"] == true -> field.name
+            else -> getAliasedFieldName(field, context, pydanticVersion)
+        }
+        2 -> when {
+            config["populate_by_name"] == true -> field.name
             else -> getAliasedFieldName(field, context, pydanticVersion)
         }
         else -> when {