+++ src/com/koxudaxi/pydantic/PydanticInspection.kt
@@ -6,6 +6,7 @@ import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import com.jetbrains.python.PyNames
 import com.jetbrains.python.codeInsight.stdlib.PyDataclassTypeProvider
+import com.jetbrains.python.codeInsight.typing.PyTypingTypeProvider
 import com.jetbrains.python.inspections.PyInspection
 import com.jetbrains.python.inspections.PyInspectionVisitor
 import com.jetbrains.python.inspections.quickfix.RenameParameterQuickFix
@@ -36,7 +37,8 @@ class PydanticInspection : PyInspection() {
             super.visitPyFunction(node)
 
             if (getPydanticModelByAttribute(node, true, myTypeEvalContext) == null) return
-            if (!node.isValidatorMethod(pydanticCacheService.getOrPutVersion())) return
+            if (!node.hasValidatorMethod(pydanticCacheService.getOrPutVersion())) return
+            if (node.hasModelValidatorModeAfter()) return
             val paramList = node.parameterList
             val params = paramList.parameters
             val firstParam = params.firstOrNull()
@@ -83,14 +85,13 @@ class PydanticInspection : PyInspection() {
             super.visitPyTypeDeclarationStatement(node)
 
             inspectAnnotatedField(node)
+            inspectCustomRootField(node)
         }
 
         override fun visitPyClass(node: PyClass) {
             super.visitPyClass(node)
 
-            if(pydanticCacheService.isV2) {
-                inspectCustomRootFieldV2(node)
-            }
+
             inspectConfig(node)
             inspectDefaultFactory(node)
         }
@@ -98,17 +99,15 @@ class PydanticInspection : PyInspection() {
         private fun inspectValidatorField(pyStringLiteralExpression: PyStringLiteralExpression) {
             if (pyStringLiteralExpression.reference?.resolve() != null) return
             val pyArgumentList = pyStringLiteralExpression.parent as? PyArgumentList ?: return
-            pyArgumentList.getKeywordArgument("check_fields")?.let { it ->
-                val checkFields = when (val value = it.valueExpression){
-                   is PyReferenceExpression -> (value.reference.resolve() as? PyTargetExpression)?.findAssignedValue()
-                   else -> value
-                }?.let { PyEvaluator.evaluateAsBoolean(it) }
+            pyArgumentList.getKeywordArgument("check_fields")?.let {
                 // ignore unresolved value
-                if (checkFields != true) return
+                if (PyEvaluator.evaluateAsBoolean(it.value)!= true) return
             }
+            val stringValue = pyStringLiteralExpression.stringValue
+            if (stringValue == "*") return
             registerProblem(
                 pyStringLiteralExpression,
-                "Cannot find field '${pyStringLiteralExpression.stringValue}'",
+                "Cannot find field '${stringValue}'",
                 ProblemHighlightType.GENERIC_ERROR
             )
         }
@@ -143,16 +142,6 @@ class PydanticInspection : PyInspection() {
             pyFunction.statementList.statements.filterIsInstance<PyExpressionStatement>()
                 .mapNotNull { (it.expression as? PyCallExpression)?.getArgument(1, PyReferenceExpression::class.java) }
                 .any { (it.reference.resolve() as? PyTargetExpression)?.findAssignedValue()?.name == "PydanticDeprecatedSince20" }
-
-
-        private fun inspectCustomRootFieldV2(pyClass: PyClass) {
-            if (getRootField(pyClass) == null) return
-            if (!isPydanticModel(pyClass, false, myTypeEvalContext)) return
-            registerProblem(
-                pyClass.nameNode?.psi,
-                "__root__ models are no longer supported in v2; a migration guide will be added in the near future", ProblemHighlightType.GENERIC_ERROR
-            )
-        }
 
         private fun inspectDefaultFactory(pyClass: PyClass) {
             if (!isPydanticModel(pyClass, true, myTypeEvalContext)) return
@@ -310,16 +299,41 @@ class PydanticInspection : PyInspection() {
         }
 
         private fun inspectCustomRootField(node: PyAssignmentStatement) {
-            val pyClass = getPydanticModelByAttribute(node, false, myTypeEvalContext) ?: return
+            val field = node.leftHandSideExpression as? PyTargetExpression ?: return
+            inspectCustomRootField(field)
+        }
+        private fun inspectCustomRootField(node: PyTypeDeclarationStatement) {
+            val field = node.target as? PyTargetExpression ?: return
+            inspectCustomRootField(field)
+        }
+        private fun inspectCustomRootField(field: PyTargetExpression) {
+            val pyClass = getPydanticModelByAttribute(field.parent, false, myTypeEvalContext) ?: return
 
-            val fieldName = (node.leftHandSideExpression as? PyTargetExpressionImpl)?.text ?: return
+            if (PyTypingTypeProvider.isClassVar(field, myTypeEvalContext)) return
+            val fieldName = field.text ?: return
+            val isV2 = pydanticCacheService.isV2
+            if (isV2 && fieldName == "__root__") {
+                registerProblem(
+                    pyClass.nameNode?.psi,
+                    "__root__ models are no longer supported in v2; a migration guide will be added in the near future", ProblemHighlightType.GENERIC_ERROR
+                )
+                registerProblem(field, "To define root models, use `pydantic.RootModel` rather than a field called '__root__'", ProblemHighlightType.WARNING)
+                return
+            }
             if (fieldName.startsWith('_')) return
-            val rootModel = getRootField(pyClass)?.containingClass ?: return
-            if (!isPydanticModel(rootModel, false, myTypeEvalContext)) return
-            registerProblem(
-                node,
-                "__root__ cannot be mixed with other fields", ProblemHighlightType.WARNING
-            )
+            val message = when {
+                isV2 -> {
+                    if (fieldName == "root") return
+                    if (!isSubClassOfPydanticRootModel(pyClass, myTypeEvalContext)) return
+                    if (pyClass.findClassAttribute("root", true, myTypeEvalContext) == null) return
+                    "Unexpected field with name ${fieldName}; only 'root' is allowed as a field of a `RootModel`"
+                }
+                else -> {
+                    if (pyClass.findClassAttribute("__root__", true, myTypeEvalContext) == null) return
+                    "__root__ cannot be mixed with other fields"
+                }
+            }
+            registerProblem(field, message, ProblemHighlightType.WARNING)
         }
 
         private fun validateDefaultAndDefaultFactory(default: PyExpression?, defaultFactory: PyExpression?): Boolean {
@@ -391,9 +405,6 @@ class PydanticInspection : PyInspection() {
             }
         }
 
-        private fun getRootField(pyClass: PyClass): PyTargetExpression? {
-            return pyClass.findClassAttribute("__root__", true, myTypeEvalContext)
-        }
     }
 
 //    override fun createOptionsPanel(): JComponent? {