+++ testData/inspection/customRoot.py
@@ -1,3 +1,5 @@
+from typing import ClassVar
+
 from pydantic import BaseModel
 
 
@@ -11,7 +13,7 @@ class B(BaseModel):
 
 class C(BaseModel):
     __root__ = 'xyz'
-    <warning descr="__root__ cannot be mixed with other fields">b = 'xyz'</warning>
+    <warning descr="__root__ cannot be mixed with other fields">b</warning> = 'xyz'
 
 
 class D(BaseModel):
@@ -26,4 +28,12 @@ class E:
 def f():
     __root__ = 'xyz'
     g = 'xyz'
+
+class G(BaseModel):
+    ATTRIBUTE_NAME: ClassVar[str] = "testing"
+    __root__ = 'xyz'
+
+class H(BaseModel):
+    __root__ = 'xyz'
+    <warning descr="__root__ cannot be mixed with other fields">b</warning>: str