+++ testData/mock/pydanticv2/root_model.py
@@ -1,0 +1,10 @@
+import typing
+
+from .main import BaseModel
+
+RootModelRootType = typing.TypeVar('RootModelRootType')
+
+
+
+class RootModel(BaseModel, typing.Generic[RootModelRootType]):
+    root: RootModelRootType