+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [0.4.7] - 2023-08-07
+- Fix false positive "__root__ cannot be mixed with other fields" -- ty… [#780](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/780)]
+- treat model validator as an instance method when mode='after' [#779](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/779)]
+- Fix NullPointerException of getExternalDocumentationUrl [#778](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/778)]
+- Fix false positive validator asterisk [[#777](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/777)]
+
 ## [0.4.6-232] - 2023-07-20
 - Associating fields with validator fields for 232 [[#762](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/762)]
 - Add validator field argument completion for 232 [[#758](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull/758)]
@@ -426,7 +432,8 @@
 - Support ellipsis(...) in fields [[#34](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull//34)]
 - Support Schema in fields [[#31](https://github.com/koxudaxi/pydantic-pycharm-plugin/pull//31)]
 
-[Unreleased]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.6-232...HEAD
+[Unreleased]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.7...HEAD
+[0.4.7]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.6-232...v0.4.7
 [0.4.6-232]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.5-232...v0.4.6-232
 [0.4.5-232]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.4-232...v0.4.5-232
 [0.4.4-232]: https://github.com/koxudaxi/pydantic-pycharm-plugin/compare/v0.4.3-232...v0.4.4-232